Move allow list to state keeper
