+++ core/lib/config/src/configs/api.rs
@@ -220,9 +220,6 @@ pub struct Web3JsonRpcConfig {
     /// (hundreds or thousands RPS).
     #[serde(default)]
     pub extended_api_tracing: bool,
-    /// Configuration options for the deployment allow list
-    #[serde(default)]
-    pub deployment_allowlist: Option<DeploymentAllowlist>,
 }
 
 impl Web3JsonRpcConfig {
@@ -261,7 +258,6 @@ impl Web3JsonRpcConfig {
             whitelisted_tokens_for_aa: vec![],
             api_namespaces: None,
             extended_api_tracing: false,
-            deployment_allowlist: None,
         }
     }
 
@@ -403,42 +399,6 @@ pub struct MerkleTreeApiConfig {
 impl MerkleTreeApiConfig {
     const fn default_port() -> u16 {
         3_072
-    }
-}
-
-#[derive(Debug, Deserialize, Clone, PartialEq)]
-pub enum DeploymentAllowlist {
-    Dynamic(DeploymentAllowlistDynamic),
-    Static(Vec<Address>),
-}
-
-#[derive(Default, Debug, Deserialize, Clone, PartialEq)]
-pub struct DeploymentAllowlistDynamic {
-    /// If `Some(url)`, allowlisting is enabled. If `None`, it's disabled.
-    /// If the `String` is empty, treat it as invalid and effectively disable.
-    http_file_url: Option<String>,
-    /// Private field for the refresh interval (in seconds).
-    refresh_interval_secs: Option<u64>,
-}
-
-impl DeploymentAllowlistDynamic {
-    /// Create a new `DeploymentAllowlist` instance.
-    pub fn new(http_file_url: Option<String>, refresh_interval_secs: Option<u64>) -> Self {
-        Self {
-            http_file_url,
-            refresh_interval_secs,
-        }
-    }
-
-    /// Returns the allowlist file URL, if present and non-empty.
-    pub fn http_file_url(&self) -> Option<&str> {
-        self.http_file_url.as_deref().filter(|s| !s.is_empty())
-    }
-
-    /// Returns the refresh interval used to reload the allowlist.
-    /// Defaults to 5 minutes if not set.
-    pub fn refresh_interval(&self) -> Duration {
-        Duration::from_secs(self.refresh_interval_secs.unwrap_or(300))
     }
 }
